+++ src/__tests__/unit/openAI/core/assistant.unit.test.ts
@@ -14,8 +14,9 @@ describe('Assistant', () => {
 
   it('Should be possible to create an Assistant', async () => {
     const chatRepository = new ChatRepository();
+    const messageRepository = new MessageRepository();
 
-    const assistant = new Assistant(chatRepository);
+    const assistant = new Assistant(chatRepository, messageRepository);
     expect(assistant.context.length).toBe(2);
     expect(assistant.humor).toBe(EHumor.SARCASTIC);
     expect(assistant.model).toBe(EModel['GPT-3.5-TURBO']);
@@ -27,47 +28,50 @@ describe('Assistant', () => {
 
   it('Should be possible to send message to GPT and store the response into the right chat messages', async () => {
     const chatRepository = new ChatRepository();
-    const assistant = new Assistant(chatRepository);
+    const messageRepository = new MessageRepository();
+
+    const assistant = new Assistant(chatRepository, messageRepository);
     const openAIApi = assistant.getOpenAIApi();
 
     (jest.spyOn(openAIApi, 'createChatCompletion') as any).mockImplementation(
       createChatCompletionMock,
     );
 
-    const messageRepository = new MessageRepository();
-
-    const chat = new Chat(
-      {
-        _id: 'My_Chat_ID',
-        ownerId: 'Almera',
-        title: 'My Subject',
-        description: 'My Description',
-      },
-      messageRepository,
-    );
+    const chat = new Chat({
+      _id: 'My_Chat_ID',
+      ownerId: 'Almera',
+      title: 'My Subject',
+      description: 'My Description',
+    });
     assistant.addChat({ chat });
 
     const myChat = (await assistant.getChat({ chatId: chat.id })) as Chat;
 
-    myChat.addMessage({
+    assistant.addMessage({
       content: 'content 01',
       role: 'user',
       ownerId: myChat.ownerId,
+      chatId: myChat.id,
     });
-    myChat.addMessage({
+    assistant.addMessage({
       content: 'content 02',
       role: 'user',
       ownerId: myChat.ownerId,
+      chatId: myChat.id,
     });
-    myChat.addMessage({
+    assistant.addMessage({
       content: 'content 03',
       role: 'user',
       ownerId: myChat.ownerId,
+      chatId: myChat.id,
     });
 
-    await assistant.sendMessage({ chatId: chat.id });
+    await assistant.sendChat(chat.id);
 
-    const messages = await myChat.getMessages();
+    const messages = await assistant.getMessages({
+      chatId: myChat.id,
+      ownerId: myChat.ownerId,
+    });
 
     expect(messages.length).toBeGreaterThan(3);
 
@@ -87,27 +91,21 @@ describe('Assistant', () => {
   it('Should be possible add different chats', async () => {
     const chatRepository = new ChatRepository();
     const messageRepository = new MessageRepository();
-    const chat = new Chat(
-      {
-        _id: 'My_Chat_ID_01',
-        ownerId: 'Almera',
-        title: 'My Subject Almera',
-        description: 'My Description Almera',
-      },
-      messageRepository,
-    );
+    const chat = new Chat({
+      _id: 'My_Chat_ID_01',
+      ownerId: 'Almera',
+      title: 'My Subject Almera',
+      description: 'My Description Almera',
+    });
 
-    const chat2 = new Chat(
-      {
-        _id: 'My_Chat_ID_02',
-        ownerId: 'Andre',
-        title: 'My Subject Andre',
-        description: 'My Description Andre',
-      },
-      messageRepository,
-    );
+    const chat2 = new Chat({
+      _id: 'My_Chat_ID_02',
+      ownerId: 'Andre',
+      title: 'My Subject Andre',
+      description: 'My Description Andre',
+    });
 
-    const assistant = new Assistant(chatRepository);
+    const assistant = new Assistant(chatRepository, messageRepository);
     assistant.addChat({ chat });
     assistant.addChat({ chat: chat2 });
 
@@ -123,47 +121,35 @@ describe('Assistant', () => {
   it('Should be possible get chats paginated', async () => {
     const chatRepository = new ChatRepository();
     const messageRepository = new MessageRepository();
-    const chat = new Chat(
-      {
-        _id: 'My_Chat_ID_01',
-        ownerId: 'Almera',
-        title: 'My Subject Almera',
-        description: 'My Description Almera',
-      },
-      messageRepository,
-    );
+    const chat = new Chat({
+      _id: 'My_Chat_ID_01',
+      ownerId: 'Almera',
+      title: 'My Subject Almera',
+      description: 'My Description Almera',
+    });
 
-    const chat2 = new Chat(
-      {
-        _id: 'My_Chat_ID_02',
-        ownerId: 'Almera',
-        title: 'My Subject Almera 02',
-        description: 'My Description Almera 02',
-      },
-      messageRepository,
-    );
+    const chat2 = new Chat({
+      _id: 'My_Chat_ID_02',
+      ownerId: 'Almera',
+      title: 'My Subject Almera 02',
+      description: 'My Description Almera 02',
+    });
 
-    const chat3 = new Chat(
-      {
-        _id: 'My_Chat_ID_03',
-        ownerId: 'Almera',
-        title: 'My Subject Almera 03',
-        description: 'My Description Almera 03',
-      },
-      messageRepository,
-    );
+    const chat3 = new Chat({
+      _id: 'My_Chat_ID_03',
+      ownerId: 'Almera',
+      title: 'My Subject Almera 03',
+      description: 'My Description Almera 03',
+    });
 
-    const otherUserChat = new Chat(
-      {
-        _id: 'My_Chat_ID_Andre',
-        ownerId: 'Andre',
-        title: 'My Subject Andre 02',
-        description: 'My Description Andre 02',
-      },
-      messageRepository,
-    );
+    const otherUserChat = new Chat({
+      _id: 'My_Chat_ID_Andre',
+      ownerId: 'Andre',
+      title: 'My Subject Andre 02',
+      description: 'My Description Andre 02',
+    });
 
-    const assistant = new Assistant(chatRepository);
+    const assistant = new Assistant(chatRepository, messageRepository);
     assistant.addChat({ chat });
     assistant.addChat({ chat: chat2 });
     assistant.addChat({ chat: otherUserChat });
@@ -192,6 +178,6 @@ describe('Assistant', () => {
     expect(chatsAndre.length).toBe(1);
     expect(chatsAndreSkipped.length).toBe(0);
     expect(chatsAlmeraSkipped.length).toBe(1);
-    expect(chatsAlmeraSkipped[0]._id).toBe('My_Chat_ID_02');
+    expect(chatsAlmeraSkipped[0].id).toBe('My_Chat_ID_02');
   });
 });